+++ packages/ivy-design-wc/src/components/date-picker/src/index.vue
@@ -6,7 +6,7 @@ import { useHost } from '@/hooks/useHostElement'
 import { CloseIcon as Close } from '@/utils/icons'
 
 defineOptions({
-    name: 'TimeSelect',
+    name: 'DatePicker',
     inheritAttrs: false
 })
 
@@ -164,17 +164,19 @@ onBeforeUnmount(() => {
     <transition name="dropdown">
         <div class="select-option-wrap" ref="targetRef" v-show="visible">
             <div class="select-arrow" ref="arrowRef"></div>
-            <ivy-scrollbar class="select-option-scroll" max-height="274px" @click="handlerClick">
-                <div
-                    v-for="item in dateList"
-                    :key="item.value"
-                    :data-value="item.value"
-                    class="option-item"
-                    :data-disabled="item.disabled === true"
-                >
-                    {{ item.value }}
+            <div class="select-option-scroll">
+                <div class="select-option" @click="handlerClick">
+                    <div
+                        v-for="item in dateList"
+                        :key="item.value"
+                        :data-value="item.value"
+                        class="option-item"
+                        :data-disabled="item.disabled === true"
+                    >
+                        {{ item.value }}
+                    </div>
                 </div>
-            </ivy-scrollbar>
+            </div>
         </div>
     </transition>
 </template>
@@ -261,11 +263,13 @@ onBeforeUnmount(() => {
     border-radius: 2px;
     overflow: hidden;
     z-index: 10;
-    padding: 6px 0;
 }
 
 .select-option-scroll {
+    overflow-x: hidden;
+    overflow-y: auto;
     background-color: #fff;
+    max-height: 274px;
     border-radius: 4px;
     border: 1px solid var(--ivy-border-color, #dcdfe6);
     box-shadow: var(
@@ -273,8 +277,24 @@ onBeforeUnmount(() => {
         0px 12px 32px 4px rgba(0, 0, 0, 0.04),
         0px 8px 20px rgba(0, 0, 0, 0.08)
     );
+    &::-webkit-scrollbar {
+        width: 6px;
+        height: 6px;
+        background-color: #fff;
+        border-radius: 2px;
+    }
+    &::-webkit-scrollbar-thumb {
+        background-color: rgba(144, 147, 153, 1);
+        border-radius: 2px;
+    }
+    &::-webkit-scrollbar-track {
+        background-color: rgba(144, 147, 153, 0.3);
+        border-radius: 2px;
+    }
 }
-
+.select-option {
+    margin: 6px 0;
+}
 .select-arrow {
     height: 6px;
     width: 6px;