+++ packages/ivy-design-wc/src/components/tree/child.tsx
@@ -1,10 +1,10 @@
-import { type SetupContext } from 'vue'
-import CollapseTransition from '@/utils/collapse-transition'
+import { type SetupContext, Transition } from 'vue'
+import { transition } from '@/utils/collapse-transition'
 import { CaretRight } from '@/utils/icons'
 
 const TreeItem = (props: Record<string, any>, { emit }: SetupContext) => {
     const { data, props: childProps } = props
-
+    const transitionAttrs = transition()
     const handleClick = () => {
         data.isOpen = !data.isOpen
         emit('item-click', data)
@@ -23,13 +23,22 @@ const TreeItem = (props: Record<string, any>, { emit }: SetupContext) => {
                 <span>{data[childProps.label]}</span>
             </div>
             {data.isLeaf ? null : (
-                <CollapseTransition>
+                <Transition
+                    onBeforeEnter={transitionAttrs.onBeforeEnter}
+                    onEnter={transitionAttrs.onEnter}
+                    onAfterEnter={transitionAttrs.onAfterEnter}
+                    onBeforeLeave={transitionAttrs.onBeforeLeave}
+                    onLeave={transitionAttrs.onLeave}
+                    onAfterLeave={transitionAttrs.onAfterLeave}
+                >
                     <div class="tree-item__children" v-show={data.isOpen}>
-                        {data[childProps.children].map((c: any) => (
-                            <TreeItem data={c} props={childProps}></TreeItem>
-                        ))}
+                        <div>
+                            {data[childProps.children].map((c: any) => (
+                                <TreeItem data={c} props={childProps}></TreeItem>
+                            ))}
+                        </div>
                     </div>
-                </CollapseTransition>
+                </Transition>
             )}
         </div>
     )